Backend project updated based on the latest DB schema.

+++ springsecurity6/springsecurity/src/main/java/com/example/repository/CustomerRepository.java
@@ -2,9 +2,10 @@ package com.example.repository;
 
 import com.example.model.Customer;
 import org.springframework.data.repository.CrudRepository;
-
 import java.util.List;
 
-public interface CustomerRepository extends CrudRepository<Customer, Integer> {
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
+
     List<Customer> findByEmail(String email);
+    
 }
